Move the console package under executor
Currently, only executor uses the `console` package.
