+++ pkg/internal/checkup/executor/console/console.go
